i18n_translation/requirements.txt
python-dotenv==0.19.0
pyspellchecker
